e2e/pyproject.toml
[project]
name = "hr-prescan-e2e"
version = "0.1.0"
description = "E2E tests for HR PreScan platform"
requires-python = ">=3.12"
dependencies = [
    "pytest>=8.0",
    "pytest-playwright>=0.6",
    "pytest-xdist>=3.6",
    "python-dotenv>=1.0",
    "requests>=2.31",
]

[tool.pytest.ini_options]
testpaths = ["api", "ui", "llm"]
# Skip LLM tests by default — they hit paid LLM APIs and are slow. Run them
# explicitly with `pytest -m llm`.
addopts = "-v --tb=short -m 'not llm'"
markers = [
    "api: API-level tests (no browser)",
    "ui: Browser UI tests",
    "llm: Tests that exercise the prescan AI end-to-end (slow, costs credits)",
    "mobile: UI tests that run at a mobile viewport",
    "social_auth: Requires ALLOW_E2E_HOOKS=true on the target server",
]
